feat: разбил на приложение на два. Собирается, запускается.

+++ gateway/src/main/java/ru/practicum/shareit/item/ItemClient.java
@@ -25,7 +25,7 @@ public class ItemClient extends BaseClient {
         );
     }
 
-    public ResponseEntity<Object> getAll(Long userId, Long from, Long size) {
+    public ResponseEntity<Object> getAll(Long userId, Integer from, Integer size) {
         Map<String, Object> parameters = Map.of(
                 "from", from,
                 "size", size
